third_party/WebKit/LayoutTests/bluetooth/requestDevice/single-filter-two-services-succeeds.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('GlucoseHeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['glucose', 'tx_power']}]
    })).then(device => assert_equals(device.name, 'Glucose Device'));
}, 'Filter with 2 services returns a matching device.');
</script>
